+++ test/unit/async-init.js
@@ -17,7 +17,7 @@
 
 import '../'
 import { describe, it } from 'mocha'
-import AsyncInit from '../../es/utils/async-init'
+import AsyncInit from '../../src/utils/async-init'
 
 describe('AsyncInit', function () {
   it('composes Stamps', async () => {
