+++ test/unit/crypto.js
@@ -18,7 +18,7 @@
 import '../'
 import { describe, it } from 'mocha'
 import { assert, expect } from 'chai'
-import * as Crypto from '../../es/utils/crypto'
+import * as Crypto from '../../src/utils/crypto'
 
 // These keys are fixations for the encryption lifecycle tests and will
 // not be used for signing
